+++ src/rules.rs
@@ -15,7 +15,6 @@ pub mod ban_unknown_rule_code;
 pub mod ban_untagged_ignore;
 pub mod ban_untagged_todo;
 pub mod ban_unused_ignore;
-pub mod button_has_type;
 pub mod camelcase;
 pub mod constructor_super;
 pub mod default_param_last;
@@ -28,6 +27,7 @@ pub mod fresh_server_event_handlers;
 pub mod getter_return;
 pub mod guard_for_in;
 pub mod jsx_boolean_value;
+pub mod jsx_button_has_type;
 pub mod jsx_curly_braces;
 pub mod jsx_key;
 pub mod jsx_no_children_prop;
@@ -256,7 +256,7 @@ fn get_all_rules_raw() -> Vec<Box<dyn LintRule>> {
     Box::new(ban_untagged_ignore::BanUntaggedIgnore),
     Box::new(ban_untagged_todo::BanUntaggedTodo),
     Box::new(ban_unused_ignore::BanUnusedIgnore),
-    Box::new(button_has_type::JSXButtonHasType),
+    Box::new(jsx_button_has_type::JSXButtonHasType),
     Box::new(camelcase::Camelcase),
     Box::new(constructor_super::ConstructorSuper),
     Box::new(default_param_last::DefaultParamLast),
